// src/common/DescriptorFields.d.ts
export default interface DescriptorFields {
    "build.version.release": string;
    "build.version.sdk": string;
    "ro.product.cpu.abi": string;
    "product.manufacturer": string;
    "product.model": string;
    "wifi.interface": string;
    udid: string;
    state: string;
    ip: string;
    pid: number;
}